[MATH] Added non square matrices 2x3,2x4,3x2,3x4,4x2,4x3 [MATH] Set vector dot function as a member function [MATH] Added a function that compute a oriented angle between 2 vectors [VIDEO] Added non square matrices uniform extensions

+++ sources/video/font_engine.cpp
@@ -100,7 +100,7 @@ namespace djah { namespace video {
 		{
 			const font_data::char_metrics &c = current_font_->metrics[*it];
 			glPushMatrix();
-			glTranslatef(x, y - c.offset.y, 0.0f);
+			glTranslatef((float)x, y - c.offset.y, 0.0f);
 			if( c.tex && c.tex->isValid() )
 			{
 				c.tex->bind();
